Feature/#19

closes #19

+++ src/los/src/main/java/at/uibk/los/model/interfaces/IScoreView.java
@@ -4,5 +4,5 @@ public interface IScoreView {
 
 	IApproach getApproach();
 	
-	int getScore();
+	double getScore();
 }
